// src/app/models/descargo.ts
export class DocDescargo {
    id_descargo?: number;
    id_sumariado?: number;
    id_usuario?: number;
    id_perfil?: number;
    ci?: string;
    id_proceso?: number;
    nomdescargo?: string;
    refdescargo?: string;
    fecha_registro?: Date;
    rutaservidor?: string;
    nombreservidor?: string;
    usuario_registro?: string;
    archivoFisico?: File;
}